Create <0xc2B5027907673041279C9006C8f83968E75A42b9>.json
DMZ is a different kind of digital currency for a different kind of person. DMZ is a cryptocurrency for people who want to live in a worldwide DeMilitarized Zone. 

Dictionary.com defines a demilitarized zone as an area in which it is forbidden to station military forces or maintain military installations. It is any area, place, or circumstance in which conflicts or hostilities are held in abeyance. Abbreviation: DMZ 

We chose the name DMZ, because it suggests a place of peace where people of different views and hues live side by side in harmony without the need to harm each other. So if you feel like we feel about creating a better world to live in and to lea